fix: cast NULL to INTEGER in V6 migration for total_installments

+++ mfc-back/src/main/resources/db/migration/V6__create_transaction_groups.sql
@@ -19,7 +19,7 @@ SELECT DISTINCT
         WHEN t.description LIKE '%[Recorrente]%' THEN 'RECURRING'
         ELSE 'INSTALLMENT'
     END,
-    NULL,
+    NULL::INTEGER,
     NOW(),
     NOW()
 FROM transactions t
